feat(package): skip installation of password manager if preinstalled

+++ .chezmoiscripts/windows/.install-password-manager.ps1
@@ -1,3 +1,8 @@
+if (Get-Command keepassxc.exe -ErrorAction SilentlyContinue) {
+  Write-Host "KeePassXC is already installed. Skipping"
+  exit
+}
+
 winget install -e --id KeePassXCTeam.KeePassXC
 function Add-ToPath {
     param (
